♻️REFACTOR: 클래스 이름 및 패키지 변경
Discount 들 앞에 Christmas 추가.
