+++ src/main/java/christmas/model/domain/event/discount/christmas/ChristmasSpecialDiscount.java
@@ -1,9 +1,10 @@
-package christmas.model.domain.event.discount;
+package christmas.model.domain.event.discount.christmas;
 
 import christmas.model.domain.date.Date;
+import christmas.model.domain.event.discount.Discount;
 import christmas.model.domain.order.RequestOrders;
 
-public class SpecialDiscount extends Discount{
+public class ChristmasSpecialDiscount extends Discount {
     private static final int SPECIAL_DISCOUNT_AMOUNT = 1000;
     private static final String SPECIAL_DISCOUNT_MESSAGE = "특별 할인: %,d원";
 
@@ -15,7 +16,7 @@ public class SpecialDiscount extends Discount{
     }
 
     @Override
-    public void reset() {
+    protected void reset() {
         resetDiscount();
     }
