+++ src/main/java/christmas/model/domain/event/discount/christmas/ChristmasWeekDiscount.java
@@ -1,11 +1,12 @@
-package christmas.model.domain.event.discount;
+package christmas.model.domain.event.discount.christmas;
 
 import static christmas.model.domain.menu.MenuCategory.DESSERT;
 
 import christmas.model.domain.date.Date;
+import christmas.model.domain.event.discount.Discount;
 import christmas.model.domain.order.RequestOrders;
 
-public class WeekDiscount extends Discount {
+public class ChristmasWeekDiscount extends Discount {
     static final int WEEK_AND_WEEKEND_EVENT_DISCOUNT_AMOUNT = 2023;
     private static final String SPECIAL_DISCOUNT_MESSAGE = "평일 할인: %,d원";
 
@@ -19,7 +20,7 @@ public class WeekDiscount extends Discount {
     }
 
     @Override
-    public void reset() {
+    protected void reset() {
         resetDiscount();
     }
