+++ src/test/java/christmas/model/domain/event/discount/ChristmasDDayDiscountTest.java
@@ -4,11 +4,12 @@ import static christmas.model.domain.order.RequestOrdersTest.makeRequestOrdersWi
 import static org.assertj.core.api.Assertions.assertThat;
 
 import christmas.model.domain.date.Date;
+import christmas.model.domain.event.discount.christmas.ChristmasDDayDiscount;
 import christmas.model.domain.order.RequestOrders;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class ChristmasDiscountTest {
+class ChristmasDDayDiscountTest {
 
     @DisplayName("3일 가지고 discount 를 계산 하면 출력값으로 '크리스마스 디데이 할인: -1,200원' 을 반환한다.")
     @Test
@@ -16,9 +17,9 @@ class ChristmasDiscountTest {
         RequestOrders requestOrders = makeRequestOrdersWithTwoMainThreeDessertThreeAppetizer();
         Date date = new Date(3);
 
-        ChristmasDiscount christmasDiscount = new ChristmasDiscount();
-        christmasDiscount.calculateDiscountAmount(requestOrders, date);
-        assertThat(christmasDiscount.toString()).isEqualTo("크리스마스 디데이 할인: -1,200원");
+        ChristmasDDayDiscount christmasDDayDiscount = new ChristmasDDayDiscount();
+        christmasDDayDiscount.calculateDiscountAmount(requestOrders, date);
+        assertThat(christmasDDayDiscount.toString()).isEqualTo("크리스마스 디데이 할인: -1,200원");
     }
 
     @DisplayName("26일 가지고 discount 를 계산 하면 할인은 없다.")
@@ -27,9 +28,9 @@ class ChristmasDiscountTest {
         RequestOrders requestOrders = makeRequestOrdersWithTwoMainThreeDessertThreeAppetizer();
         Date date = new Date(26);
 
-        ChristmasDiscount christmasDiscount = new ChristmasDiscount();
-        christmasDiscount.calculateDiscountAmount(requestOrders, date);
-        assertThat(christmasDiscount.toString()).isEqualTo("크리스마스 디데이 할인: 0원");
+        ChristmasDDayDiscount christmasDDayDiscount = new ChristmasDDayDiscount();
+        christmasDDayDiscount.calculateDiscountAmount(requestOrders, date);
+        assertThat(christmasDDayDiscount.toString()).isEqualTo("크리스마스 디데이 할인: 0원");
     }
 
 }
