+++ src/test/java/christmas/model/domain/event/discount/DiscountTest.java
@@ -4,6 +4,10 @@ import static christmas.model.domain.order.RequestOrdersTest.makeRequestOrdersWi
 import static org.assertj.core.api.Assertions.assertThat;
 
 import christmas.model.domain.date.Date;
+import christmas.model.domain.event.discount.christmas.ChristmasDDayDiscount;
+import christmas.model.domain.event.discount.christmas.ChristmasSpecialDiscount;
+import christmas.model.domain.event.discount.christmas.ChristmasWeekDiscount;
+import christmas.model.domain.event.discount.christmas.ChristmasWeekendDiscount;
 import christmas.model.domain.order.RequestOrders;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -25,7 +29,7 @@ class DiscountTest {
     @DisplayName("크리스마스 11일 전 이벤트 할인 금액을 계산하면 -2300원이 나온다.")
     @Test
     void discountTest_Christmas() {
-        Discount discount = new ChristmasDiscount();
+        Discount discount = new ChristmasDDayDiscount();
         discount.calculateDiscountAmount(requestOrders, normalDate);
 
         assertThat(discount.getDiscountAmount()).isEqualTo(-2300);
@@ -34,7 +38,7 @@ class DiscountTest {
     @DisplayName("SpecialDay 로 이벤트 할인 금액을 계산하면 -1000원이 나온다.")
     @Test
     void discountTest_Special() {
-        Discount discount = new SpecialDiscount();
+        Discount discount = new ChristmasSpecialDiscount();
         discount.calculateDiscountAmount(requestOrders, specialDate);
 
         assertThat(discount.getDiscountAmount()).isEqualTo(-1000);
@@ -43,7 +47,7 @@ class DiscountTest {
     @DisplayName("평일에 디져트 3개를 주문으로 할일 금액을 계산하면 -6069 가 나온다.")
     @Test
     void discountTest_Week() {
-        Discount discount = new WeekDiscount();
+        Discount discount = new ChristmasWeekDiscount();
         discount.calculateDiscountAmount(requestOrders, normalDate);
 
         assertThat(discount.getDiscountAmount()).isEqualTo(-6069);
@@ -52,7 +56,7 @@ class DiscountTest {
     @DisplayName("주말에 메인 2개를 주문으로 할일 금액을 계산하면 -4046 이 나온다.")
     @Test
     void discountTest_Weekend() {
-        Discount discount = new WeekendDiscount();
+        Discount discount = new ChristmasWeekendDiscount();
         discount.calculateDiscountAmount(requestOrders, weekendDate);
 
         assertThat(discount.getDiscountAmount()).isEqualTo(-4046);
